# openapi.yaml
openapi: 3.0.3
info:
  title: Membership Service API
  description: |
    OAuth 2.0 API for authenticating users and retrieving member information.

    This API implements Resource Owner Password Flow with client credentials validation,
    allowing external services to authenticate users securely and access their full profile data.

    For detailed integration guide with code examples, see API_INTEGRATION.md
  version: 1.0.0
  contact:
    name: Membership Service Support
    email: admin@example.com
servers:
- url: https://your-membership-service.com/api
  description: Production server
- url: http://localhost:8787/api
  description: Local development server
security:
- BearerAuth: []
...
paths:
...
  /admin/oauth-clients:
    get:
      tags:
      - Admin
      summary: List all OAuth clients
      description: Get a list of all registered OAuth clients (admin only)
      operationId: listOAuthClients
      responses:
        '200':
          description: List of OAuth clients
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  clients:
                    type: array
                    items:
                      $ref: '#/components/schemas/OAuthClientResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string

    post:
      tags:
      - Admin
      summary: Create OAuth client
      description: |
        Create a new OAuth client for external service integration (admin only).

        Returns client_id and client_secret. The client_secret is shown only once!
      operationId: createOAuthClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              properties:
                name:
                  type: string
                  description: Service name
                  example: Chatbot Service
                allowed_ips:
                  type: string
                  description: JSON array of allowed IP addresses/CIDRs (optional)
                  example: '["192.168.1.0/24", "10.0.0.1"]'
      responses:
        '200':
          description: OAuth client created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  client:
                    $ref: '#/components/schemas/OAuthClientCreatedResponse'
                  warning:
                    type: string
                    example: Save the client_secret securely. It will not be shown again!
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
  /admin/oauth-clients/{id}:
    put:
      tags:
      - Admin
      summary: Update OAuth client
      description: Update OAuth client details (admin only)
      operationId: updateOAuthClient
      parameters:
      - name: id
        in: path
        required: true
        description: OAuth client ID
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: New service name
                allowed_ips:
                  type: string
                  description: JSON array of allowed IP addresses/CIDRs
      responses:
        '200':
          description: Client updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string

    delete:
      tags:
      - Admin
      summary: Delete OAuth client
      description: Delete an OAuth client (admin only)
      operationId: deleteOAuthClient
      parameters:
      - name: id
        in: path
        required: true
        description: OAuth client ID
        schema:
          type: string
      responses:
        '200':
          description: Client deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
  /admin/pending-signups:
    get:
      tags:
      - Admin
      summary: List pending signups
      description: Get all pending signup requests (admin only)
      operationId: listPendingSignups
      responses:
        '200':
          description: List of pending signups
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  signups:
                    type: array
                    items:
                      $ref: '#/components/schemas/PendingSignup'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
  /admin/pending-signups/{id}/approve:
    post:
      tags:
      - Admin
      summary: Approve signup request
      description: Approve a pending signup and create user account (admin only)
      operationId: approveSignup
      parameters:
      - name: id
        in: path
        required: true
        description: Pending signup ID
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                membership_number:
                  type: string
                  description: Assign specific membership number (optional, will auto-generate if not provided)
                  example: MEM123456
      responses:
        '200':
          description: Signup approved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      membership_number:
                        type: string
                      email:
                        type: string
        '404':
          description: Signup request not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: Signup request not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
  /admin/pending-signups/{id}/reject:
    post:
      tags:
      - Admin
      summary: Reject signup request
      description: Reject a pending signup request (admin only)
      operationId: rejectSignup
      parameters:
      - name: id
        in: path
        required: true
        description: Pending signup ID
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for rejection (optional, will be sent to user)
                  example: Incomplete information provided
      responses:
        '200':
          description: Signup rejected successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '404':
          description: Signup request not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: Signup request not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /oauth/token endpoint
  schemas:
...
    OAuthClientResponse:
      type: object
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        client_id:
          type: string
          example: client_abc123def456789
        name:
          type: string
          example: Chatbot Service
        allowed_ips:
          type: string
          nullable: true
          example: '["192.168.1.0/24"]'
        created_at:
          type: string
          format: date-time
          example: '2026-02-05T10:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2026-02-05T10:00:00Z'
    OAuthClientCreatedResponse:
      type: object
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        client_id:
          type: string
          example: client_abc123def456789
        client_secret:
          type: string
          description: Client secret (shown only once!)
          example: secret_xyz789abc123def456
        name:
          type: string
          example: Chatbot Service
        allowed_ips:
          type: string
          nullable: true
          example: '["192.168.1.0/24"]'
        created_at:
          type: string
          format: date-time
          example: '2026-02-05T10:00:00Z'
...